Gameplay loop working, mostly
I've checked the code and made a few minor adjustments.

The game will now be able to reset without needing to reload the scene as a whole.

I've also updated the Main menu and added a Tutorial, Credits and the layout for the leaderboard.

The next step for me is to get the save & and load state for the scores.

Current issues:

1#
For the reset function within the JokeScore script I have been unable to reset the AI rating value to 0, as a result, every time a player restarts, the previous AI value will always be added on top, making an increasing score through resets.

2#
For the Gameplay UI, I attempted to make a semi-transparent colourfull boarder for text and selectable objects, but does not appear on screen regardless of Z order